금액의 타입을 BigDecimal에서 Long으로 변경
원화의 경우 소수점 자리가 필요하지 않음

+++ src/main/java/com/aipiabackend/order/controller/dto/OrderCreateRequest.java
@@ -5,7 +5,6 @@ import com.aipiabackend.order.service.dto.OrderLineCreateCommand;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderCreateRequest(
@@ -17,7 +16,7 @@ public record OrderCreateRequest(
     List<OrderLineCreateRequest> orderLines,
 
     @NotNull(message = "주문 금액은 필수입니다")
-    BigDecimal amount
+    Long amount
 ) {
 
     /**
